feat: rating (#25)
* Add constraint for max no of attacks on a base per level

* Log attacker path

* Drop table attacker_path

* fix: Add new map layout for level if it does not exist

* fix: get user from session in replay route

+++ src/api/game/util.rs
@@ -22,7 +22,7 @@ pub struct LeaderboardResponse {
 #[derive(Queryable, Deserialize, Serialize)]
 pub struct LeaderboardEntry {
     pub username: String,
-    pub overall_rating: i32,
+    pub overall_rating: f32,
 }
 
 pub fn get_leaderboard(page: i64, limit: i64, conn: &PgConnection) -> Result<LeaderboardResponse> {
